move dbapi tests into standalone file
