+++ tests/system/test_system.py
@@ -15,11 +15,9 @@
 import collections
 import datetime
 import decimal
-import hashlib
 import math
 import operator
 import os
-import pickle
 import struct
 import threading
 import time
@@ -42,8 +40,6 @@ from google.cloud.spanner_v1 import KeyRange
 from google.cloud.spanner_v1 import KeySet
 from google.cloud.spanner_v1 import BurstyPool
 from google.cloud.spanner_v1 import COMMIT_TIMESTAMP
-
-from google.cloud.spanner_dbapi.connection import Connection
 
 from test_utils.retry import RetryErrors
 from test_utils.retry import RetryInstanceState
@@ -2745,235 +2741,3 @@ class FauxCall(object):
 
     def details(self):
         return self._details
-
-
-class TestTransactionsManagement(unittest.TestCase):
-    """Transactions management support tests."""
-
-    DATABASE_NAME = "db-api-transactions-management"
-
-    DDL_STATEMENTS = (
-        """CREATE TABLE contacts (
-            contact_id INT64,
-            first_name STRING(1024),
-            last_name STRING(1024),
-            email STRING(1024)
-        )
-        PRIMARY KEY (contact_id)""",
-    )
-
-    @classmethod
-    def setUpClass(cls):
-        """Create a test database."""
-        cls._db = Config.INSTANCE.database(
-            cls.DATABASE_NAME,
-            ddl_statements=cls.DDL_STATEMENTS,
-            pool=BurstyPool(labels={"testcase": "database_api"}),
-        )
-        cls._db.create().result(30)  # raises on failure / timeout.
-
-    @classmethod
-    def tearDownClass(cls):
-        """Delete the test database."""
-        cls._db.drop()
-
-    def tearDown(self):
-        """Clear the test table after every test."""
-        self._db.run_in_transaction(clear_table)
-
-    def test_commit(self):
-        """Test committing a transaction with several statements."""
-        want_row = (
-            1,
-            "updated-first-name",
-            "last-name",
-            "test.email_updated@domen.ru",
-        )
-        # connect to the test database
-        conn = Connection(Config.INSTANCE, self._db)
-        cursor = conn.cursor()
-
-        # execute several DML statements within one transaction
-        cursor.execute(
-            """
-INSERT INTO contacts (contact_id, first_name, last_name, email)
-VALUES (1, 'first-name', 'last-name', 'test.email@domen.ru')
-        """
-        )
-        cursor.execute(
-            """
-UPDATE contacts
-SET first_name = 'updated-first-name'
-WHERE first_name = 'first-name'
-"""
-        )
-        cursor.execute(
-            """
-UPDATE contacts
-SET email = 'test.email_updated@domen.ru'
-WHERE email = 'test.email@domen.ru'
-"""
-        )
-        conn.commit()
-
-        # read the resulting data from the database
-        cursor.execute("SELECT * FROM contacts")
-        got_rows = cursor.fetchall()
-        conn.commit()
-
-        self.assertEqual(got_rows, [want_row])
-
-        cursor.close()
-        conn.close()
-
-    def test_rollback(self):
-        """Test rollbacking a transaction with several statements."""
-        want_row = (2, "first-name", "last-name", "test.email@domen.ru")
-        # connect to the test database
-        conn = Connection(Config.INSTANCE, self._db)
-        cursor = conn.cursor()
-
-        cursor.execute(
-            """
-INSERT INTO contacts (contact_id, first_name, last_name, email)
-VALUES (2, 'first-name', 'last-name', 'test.email@domen.ru')
-        """
-        )
-        conn.commit()
-
-        # execute several DMLs with one transaction
-        cursor.execute(
-            """
-UPDATE contacts
-SET first_name = 'updated-first-name'
-WHERE first_name = 'first-name'
-"""
-        )
-        cursor.execute(
-            """
-UPDATE contacts
-SET email = 'test.email_updated@domen.ru'
-WHERE email = 'test.email@domen.ru'
-"""
-        )
-        conn.rollback()
-
-        # read the resulting data from the database
-        cursor.execute("SELECT * FROM contacts")
-        got_rows = cursor.fetchall()
-        conn.commit()
-
-        self.assertEqual(got_rows, [want_row])
-
-        cursor.close()
-        conn.close()
-
-    def test_autocommit_mode_change(self):
-        """Test auto committing a transaction on `autocommit` mode change."""
-        want_row = (
-            2,
-            "updated-first-name",
-            "last-name",
-            "test.email@domen.ru",
-        )
-        # connect to the test database
-        conn = Connection(Config.INSTANCE, self._db)
-        cursor = conn.cursor()
-
-        cursor.execute(
-            """
-INSERT INTO contacts (contact_id, first_name, last_name, email)
-VALUES (2, 'first-name', 'last-name', 'test.email@domen.ru')
-        """
-        )
-        cursor.execute(
-            """
-UPDATE contacts
-SET first_name = 'updated-first-name'
-WHERE first_name = 'first-name'
-"""
-        )
-        conn.autocommit = True
-
-        # read the resulting data from the database
-        cursor.execute("SELECT * FROM contacts")
-        got_rows = cursor.fetchall()
-
-        self.assertEqual(got_rows, [want_row])
-
-        cursor.close()
-        conn.close()
-
-    def test_rollback_on_connection_closing(self):
-        """
-        When closing a connection all the pending transactions
-        must be rollbacked. Testing if it's working this way.
-        """
-        want_row = (1, "first-name", "last-name", "test.email@domen.ru")
-        # connect to the test database
-        conn = Connection(Config.INSTANCE, self._db)
-        cursor = conn.cursor()
-
-        cursor.execute(
-            """
-INSERT INTO contacts (contact_id, first_name, last_name, email)
-VALUES (1, 'first-name', 'last-name', 'test.email@domen.ru')
-        """
-        )
-        conn.commit()
-
-        cursor.execute(
-            """
-UPDATE contacts
-SET first_name = 'updated-first-name'
-WHERE first_name = 'first-name'
-"""
-        )
-        conn.close()
-
-        # connect again, as the previous connection is no-op after closing
-        conn = Connection(Config.INSTANCE, self._db)
-        cursor = conn.cursor()
-
-        # read the resulting data from the database
-        cursor.execute("SELECT * FROM contacts")
-        got_rows = cursor.fetchall()
-        conn.commit()
-
-        self.assertEqual(got_rows, [want_row])
-
-        cursor.close()
-        conn.close()
-
-    def test_results_checksum(self):
-        """Test that results checksum is calculated properly."""
-        conn = Connection(Config.INSTANCE, self._db)
-        cursor = conn.cursor()
-
-        cursor.execute(
-            """
-INSERT INTO contacts (contact_id, first_name, last_name, email)
-VALUES
-    (1, 'first-name', 'last-name', 'test.email@domen.ru'),
-    (2, 'first-name2', 'last-name2', 'test.email2@domen.ru')
-        """
-        )
-        self.assertEqual(len(conn._statements), 1)
-        conn.commit()
-
-        cursor.execute("SELECT * FROM contacts")
-        got_rows = cursor.fetchall()
-
-        self.assertEqual(len(conn._statements), 1)
-        conn.commit()
-
-        checksum = hashlib.sha256()
-        checksum.update(pickle.dumps(got_rows[0]))
-        checksum.update(pickle.dumps(got_rows[1]))
-
-        self.assertEqual(cursor._checksum.checksum.digest(), checksum.digest())
-
-
-def clear_table(transaction):
-    """Clear the test table."""
-    transaction.execute_update("DELETE FROM contacts WHERE true")
